feat: can now embed basic inline images in the image :boom:
Added support for images to the paragraph runs and to the document.
Implemented various xml structs required for this basic example to work
though there is a lot more that could go into it to make it more robust.
This represents a minimal working product / proof-of-concept for the
feature. I think I can close the issue :)

Closes #1